refactor: move generate_h1_h2_N_tilde to tss-core
